move contextmanagers to contexts
